requirements.txt
# Unified Flask API with MCP Support and WebSocket
Flask>=2.2.3
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
eventlet>=0.33.0
pythainlp>=3.1.1
requests>=2.19.1
mimerender>=0.6.0
